+++ pkg/vm_service/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 8.2.0-dev
 - Update to version `3.56` of the spec.
 - Added optional `line` and `column` properties to `SourceLocation`.
+- Added a new `SourceReportKind`, `BranchCoverage`, which reports branch level
+  coverage information.
 
 ## 8.1.0
 - Update to version `3.55` of the spec.